2013-01-24	Daniel Hilst	<danielhilst@gmail.com>
* event.c riemann_event_strfevent(): new function. It gets an
riemann_event_t event and a format string an return a formated string,
like strftime.  Here is the format specification:

Flag 	Field
%T 	time
%S	state
%s	service
%h	host
%d	description
%G	tags (Comma separated list)
%t	ttl
%a	attributes (Comma seaparated list at format `key=value')
%mi 	metric_sint64
%md	metric_d
%mf	metric_f

For scalar values, if has_FIELD is false, the flag is ignored. For
strings a (null) is inserted, this string behavior may change in
future.  Unknown flags are silenty ignored (fix this).

* exmp02.c third parameter added, format string passed to
riemann_event_strfevent()